requirements.txt
Flask==3.0.2
APScheduler==3.10.4
requests==2.31.0
python-dotenv==1.0.1
google-auth-oauthlib==1.2.0
google-api-python-client==2.120.0
Werkzeug==3.0.1
pytz==2024.1
psutil==5.9.8
Flask-Session==0.6.0
Flask-Login==0.6.3
numpy==1.26.4 